Updating README to pass my documentation standards.
While I am not updating this as I do not have any new findings or anything of substance to add. The documentation was very cumbersome and so I felt like tidying it up in passing. I have spent a lot of time in markdown so this was an eye sore. Sorry anyone who saw it in the prior state. And I hope this truly helped the people who starred it and they weren't just being nice 🥺 Data loss genuinely breaks my heart.